platform/python-sidecar/pipeline/requirements.txt
google-cloud-storage>=2.14.0,<3
google-cloud-aiplatform>=1.42.0,<2
psycopg[binary,pool]>=3.1.18,<4
pydantic>=2.6.0,<3
PyYAML>=6.0,<7
python-frontmatter>=1.1.0,<2
python-dotenv>=1.0.0,<2
tenacity>=8.2.0,<9
structlog>=24.1.0,<25
tiktoken>=0.7.0
pgvector>=0.3.0
numpy>=1.26.0
